+++ vite.config.ts
@@ -25,6 +25,7 @@ export default defineConfig(async ({ mode }) => ({
         main: path.resolve(__dirname, "index.html"),
         settings: path.resolve(__dirname, "settings.html"),
         debug: path.resolve(__dirname, "debug.html"),
+        float: path.resolve(__dirname, "float.html"),
       },
       output: {
         manualChunks(id: string) {